refactor(bayn): modularize mutation store
