+++ services/bayn/src/execution/mutations/postgres/shared.ts
@@ -1,0 +1,4 @@
+import { Effect, Result } from 'effect'
+
+export const fromDecision = <A, E>(evaluate: () => Result.Result<A, E>): Effect.Effect<A, E> =>
+  Effect.suspend(() => Effect.fromResult(evaluate()))
